smart_event_manager/requirements.txt
rich
openpyxl
python-dotenv
